\name{anolis.tree}
\alias{anolis.tree}
\docType{data}
\title{anolis.tree}
\description{Anolis sample tree from phytools package.}
\usage{data("anolis.tree")}
\format{
  A phylo object.
  }
}
\details{Anole genus}
\source{Liam J. Revell}
\references{}
\examples{
data(historical)
## maybe str(historical) ; plot(historical) ...
}
\keyword{datasets}
