
\name{heat_shock_response}
\alias{heat_shock_response}
\docType{data}
\title{Heat Shock Response}
\description{
  A pseudodataset for heat shock protein (HSP) expression levels in mammalian cells under control and heat-stressed conditions.
}
\usage{
  data("heat_shock_response")
}
\format{
  A data frame with 100 observations on the following 2 variables:
  \describe{
    \item{condition}{a factor indicating the experimental condition: 'control' or 'heat_stressed'}
    \item{HSP_expression}{a numeric vector representing the expression levels of heat shock proteins}
  }
}
\details{
  %% Simulated data for the purpose of statistical analysis and hypothesis testing.
}
\source{
  Simulated data for the heat shock protein expression levels in control and heat-stressed cells.
}
\references{}
\examples{
  data(heat_shock_response)
  ## maybe str(heat_shock_response) ; plot(heat_shock_response) ...
}
\keyword{datasets}

