\name{historical}
\alias{historical}
\docType{data}
\title{
%%   historical
}
\description{
%% A pseudodataset for historical plant sex counts.
}
\usage{data("historical")}
\format{
  A data frame with 0 observations on the following 2 variables.
  \describe{
    \item{\code{x}}{a numeric vector}
    \item{\code{y}}{a numeric vector}
  }
}
\details{
%%  Artificial data. 
}
\source{
%%  Modified from arbuthnot baptism data.
}
\references{
%%  
}
\examples{
data(historical)
## maybe str(historical) ; plot(historical) ...
}
\keyword{datasets}
