\name{nycbirds}
\alias{nycbirds}
\docType{data}
\title{nycbirds}
\description{A pseudodataset for migratory birds departing around the NY, NY area.}
\usage{data("nycbirds")}
\format{
  A data frame with 0 observations on the following 2 variables.
  \describe{
    \item{\code{x}}{a numeric vector}
    \item{\code{y}}{a numeric vector}
  }
}
\details{Artificial data.}
\source{Modified from the Bureau of Transportation Statistics flight data.}
\references{}
\examples{
data(nycbirds)
## maybe str(nycbirds) ; plot(nycbirds) ...
}
\keyword{datasets}
