\name{pollen}
\alias{pollen}
\docType{data}
\title{pollen}
\description{Pseudodata on pollen nutritional info.}
\usage{data("pollen")}
\format{
  A data frame with 0 observations on the following 2 variables.
  \describe{
    \item{\code{x}}{a numeric vector}
    \item{\code{y}}{a numeric vector}
  }
}
\details{Data is artifical.}
\source{A modified version of the fastfood dataset.}
\references{}
\examples{
data(pollen)
## maybe str(pollen) ; plot(pollen) ...
}
\keyword{datasets}
