\name{present}
\alias{present}
\docType{data}
\title{present}
\description{A pseudodataset for present plant sex counts.}
\usage{data("present")}
\format{
  A data frame with 0 observations on the following 2 variables.
  \describe{
    \item{\code{x}}{a numeric vector}
    \item{\code{y}}{a numeric vector}
  }
}
\details{
%%  Artificial data. 
}
\source{Modified from arbuthnot baptism data.}
\references{}
\examples{
data(present)
## maybe str(present) ; plot(present) ...
}
\keyword{datasets}
